pyproject.toml
[tool.poetry]
name = "codeqai"
version = "0.0.1"
description = ""
authors = ["fynnfluegge <fynnfluegge@gmx.de>"]
readme = "README.md"
packages = [{ include = "codeqai" }]

[tool.poetry.dependencies]
python = "^3.9"
tiktoken = "^0.4.0"
yaspin = "^3.0.0"
pytest = "^7.4.0"
inquirer = "^3.1.3"
gitpython = "^3.1.36"
langchain = "^0.0.297"
openai = "^0.28.0"
pyyaml = "^6.0.1"
faiss-cpu = "^1.7.4"
tree-sitter = "^0.20.2"
tree-sitter-languages = "^1.7.0"

[tool.poetry.scripts]
codeqai = "codeqai.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
